Move CachedToken class to not break globals and imports
