+++ sync/__init__.py
@@ -1,5 +1,5 @@
 """Library for leveraging the power of Sync"""
 
-__version__ = "1.10.1"
+__version__ = "1.10.2"
 
 TIME_FORMAT = "%Y-%m-%dT%H:%M:%SZ"
